+++ src/Data/index.ts
@@ -1,0 +1,2 @@
+export {default as skillsData} from "./skills";
+export {default as projectsData} from  "./projects";