+++ src/Data/projects.ts
@@ -1,0 +1,51 @@
+import ChattoMatto from "../Assets/Banners/El-Chatto-Matto-banner.png";
+
+const projects: Project[] = [
+  {
+    image: "https://raw.githubusercontent.com/Beqa-Beqa/Budgetify/master/Budgetify-banner.png",
+    title: "Budgetify",
+    description: "Budgetify is a web application built by 'React', 'D3.js', 'TS' and 'CSS'. The app let's you manage your budget effectively.",
+    urls: {
+      github: "https://github.com/Beqa-Beqa/Budgetify",
+      live: "https://beqa-beqa.github.io/Budgetify/"
+    }
+  },
+  {
+    image: ChattoMatto,
+    title: "El Chatto Matto",
+    description: "El Chatto Matto is social network like application where users can find, see profiles and text each other, authorization is done with Gmail.",
+    urls: {
+      github: "https://github.com/Beqa-Beqa/El-Chatto-Matto",
+      live: "https://web-todo-6.web.app/"
+    }
+  },
+  {
+    image: "https://raw.githubusercontent.com/Beqa-Beqa/Audiophile/master/Audiophile-banner.png",
+    title: "Audiophile ",
+    description: "Welcome to Audiophile. This web application is built by 'React', 'React-router', 'TS' and 'CSS'. The idea is about ecommerce application where you can see lots of variations of audio devices.",
+    urls: {
+      github: "https://github.com/Beqa-Beqa/Audiophile",
+      live: "https://beqa-beqa.github.io/Audiophile/"
+    }
+  },
+  {
+    image: "https://raw.githubusercontent.com/Beqa-Beqa/Dev-Finder/master/devfinder.gif",
+    title: "Dev Finder",
+    description: "The project consists of one page with responsive design and optional light or dark theme mode. The page has search field that will allow you to look for a desired user - There are only test users and not the real ones. (Github API)",
+    urls: {
+      github: "https://github.com/Beqa-Beqa/Dev-Finder",
+      live: "https://beqa-beqa.github.io/Dev-Finder/"
+    }
+  },
+  {
+    image: "https://raw.githubusercontent.com/Beqa-Beqa/Splitter/master/Splitter-banner.png",
+    title: "Splitter",
+    description: "Splitter. The app lets you to split amount to pay on multiple person accurately. ",
+    urls: {
+      github: "https://github.com/Beqa-Beqa/Splitter",
+      live: "https://beqa-beqa.github.io/Splitter/"
+    }
+  },
+];
+
+export default projects;