+++ src/Data/skills.ts
@@ -1,0 +1,60 @@
+
+/*
+* Intentionally each subArray should be length of 6 or less to be properly displayed on the screen
+*/
+const skills: Skill[][] = [
+  [
+    {
+      imgUrl: "https://miro.medium.com/v2/resize:fit:720/format:webp/1*x0d41ns8PTQZz4a3VbMrBg.png",
+      text: "React.js"
+    },
+    {
+      imgUrl: "https://www.workinggears.com/images/services/yuxa9EAfarIcX9EK8Ei1JSjwRnP7e0-metaanMtdHMuanBn-.jpg",
+      text: "TS & JS"
+    },
+    {
+      imgUrl: "https://www.w3.org/html/logo/img/html5-topper.png",
+      text: "HTML5"
+    }, 
+    {
+      imgUrl: "https://www.dice.com/binaries/content/gallery/dice/insights/2020/07/Bootstrap.png",
+      text: "Bootstrap"
+    },
+    {
+      imgUrl: "https://1000logos.net/wp-content/uploads/2020/09/CSS-Logo.png",
+      text: "CSS"
+    },
+    {
+      imgUrl: "https://embed-ssl.wistia.com/deliveries/7a380179c3a5ee2d36d3b8748022ac22e040104b.jpg",
+      text: "D3.js"
+    }
+  ],
+  [
+    {
+      imgUrl: "https://miro.medium.com/v2/resize:fit:720/format:webp/1*mtsk3fQ_BRemFidhkel3dA.png",
+      text: "Git & GitHub"
+    },
+    {
+      imgUrl: "https://miro.medium.com/v2/resize:fit:640/format:webp/0*vq-JSMynSHUPXx70",
+      text: "Express.js"
+    },
+    {
+      imgUrl: "https://i.pinimg.com/736x/2f/9c/11/2f9c11f9e55efbf1791f12c06d60729b.jpg",
+      text: "Python"
+    },
+    {
+      imgUrl: "https://avangelstech.com/wp-content/uploads/2023/09/gcp-firebase.png",
+      text: "Firebase"
+    },
+    {
+      imgUrl: "https://cdn-images-1.medium.com/v2/resize:fit:408/1*1inyeCa1-NL7dyLeX1dPzQ.png",
+      text: "SQL & NoSQL"
+    },
+    {
+      imgUrl: "https://hayqsystem.am/wp-content/uploads/2024/02/jhgj-4.jpg",
+      text: "C# Fundamentals"
+    }
+  ]
+];
+
+export default skills;